+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Invalid paths no longer result in an error. Instead, ggpath will warn the user and insert an empty grob.
 * Minor adjustment to the format of lists in the documentation to avoid check problems in an upcoming R release.
+* Added new theme element `element_raster()` that allows custom images in ggplot backgrounds. (#9)
 
 # ggpath 1.0.1